+++ ddd/logic/encodage_des_notes/encodage/domain/service/encoder_notes_en_lot.py
@@ -39,7 +39,7 @@ from ddd.logic.encodage_des_notes.encodage.repository.note_etudiant import INote
 from ddd.logic.encodage_des_notes.shared_kernel.domain.model.encoder_notes_rapport import EncoderNotesRapport
 from ddd.logic.encodage_des_notes.shared_kernel.domain.service.i_inscription_examen import IInscriptionExamenTranslator
 from ddd.logic.encodage_des_notes.shared_kernel.dtos import PeriodeEncodageNotesDTO
-from ddd.logic.encodage_des_notes.shared_kernel.repository.i_report import IEncoderNotesRapportRepository
+from ddd.logic.encodage_des_notes.shared_kernel.repository.i_encoder_notes_rapport import IEncoderNotesRapportRepository
 from ddd.logic.encodage_des_notes.soumission.dtos import DesinscriptionExamenDTO
 from osis_common.ddd import interface
